LPS-63898 Rename packages
